move CommandError send code to lib_sonnetcommands
